update : some refactor
