--- src/main/java/com/github/sakakiaruka/SettingsLoad.java
@@ -1,13 +1,0 @@
-package com.github.sakakiaruka;
-
-import org.bukkit.configuration.file.FileConfiguration;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import static com.github.sakakiaruka.Sakura.DEFAULT_CONFIG;
-
-public class SettingsLoad {
-    public static int BLOCKS_LIMIT = 4000;
-
-}
