+++ src/main/java/com/github/sakakiaruka/sakura/SettingsLoad.java
@@ -1,0 +1,6 @@
+package com.github.sakakiaruka.sakura;
+
+public class SettingsLoad {
+    public static int BLOCKS_LIMIT = 4000;
+
+}
