+++ src/main/resources/plugin.yml
@@ -1,4 +1,4 @@
 name: sakura
 version: '${project.version}'
-main: com.github.sakakiaruka.Sakura
+main: com.github.sakakiaruka.sakura.Sakura
 api-version: '1.20'
